+++ gh-pages/changelog.md
@@ -26,6 +26,7 @@ New g14n.info layouts, with link Changelog page
 - Add or remove input and output pins.
 - Emit events to let integrate with other libraries.
 - Node dynamic width considered the number of pins.
+- Better docs.
 
 ### Fixed